requirements.txt
pyirsdk==1.3.2
PyAutoGUI==0.9.54
playsound==1.3.0
requests==2.32.3